+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Assemblies are now digitally signed using cosign.
+  ([#4637](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4637))
+
 * Add support for reporting full state.
   ([#4633](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4633))